Comparison with Median
Instead of the rather complicated mode of modes, I tried the analysis with each cluster's median values. The Neural Network performs even better with the top median values than with the mode-modes, but the interpretability with respect to the individual clusters is no longer present because the antiCPy function selects much more relevant features than before (usually more than 10). So there is a trade-off between the clusters' explainability and the global surrogate predictability.