Major Overhaul: Contemporary Data
I realised that the methodology for getting the data is flawed because it relies on the current S&P 500 stocks instead of the contemporary data from the 1992-2012 period. Hence, I redid the analysis with a 2012 company list and the contemporary GICS classification with only 10 sectors. Obviously, the results change because of this, but  I think the qualitative results are similar and the interpretation of the sector importances should still be feasible. Also attached the CompaniesStocks.csv as the file of the raw company stock indices so that everyone can do the renormalisation etc. on their own, if they wish to do so.