Changepoint Analysis for mean Correlation
(I accidentally uploaded this to the wrong repo)
Analysis of the 1d time ereis of mean correlations. Change points are searched in rolling windows. The window is shifted by one observation for each new window. The composite npy results are attached, too. Only 43 of the 4291 windows have to be disregarded because of NaNs/Infs which means that 99% of the data is preserved. Hence, I do not believe that any relevant information is lost by this process.