Change Point Analysis Mean Correlation
(I accidentally first uploaded this to the wrong repo) This is a Change Point analysis for the 1d Time Series of the mean correlation. CPs are searched in rolling windows. Now the window is shifted by one observation for each new window. The composite npy results are attached, too. Only 43 of the 4291 windows have to be disregarded because of NaNs/Infs which means that 99% of the data is preserved. Hence, I do not believe that any relevant information is lost by this process.