Fixed some mistakes
The quantile number q was not set as q in every part of the code, but sometimes fixed to 100. Still, for q above 8, we have predominantly negative causalities. I assume this is caused by the following fact: the more bins we have, the rarer are combinations between those bins for the two time series. Hence, the code will probably use a default value instead.