Re-Do Stepanov's analysis
Next step of this ipynb is to re-do Stepanov's drift and diffusion estimation to then later compare it to a GLE-based estimation with the chnge points as our break points. Now, this notebook shows the estimation of g^2 (squared diffusion) for one of the approx. 100 time windows for 10 binned values (and therefore 10 mean values). This has to be re-written as a function in order to be used in every singly moving window in just one big for-loop.